test(flightdeck): isolate FD emit tests into their own bun process (#464)
The two FlightDeck emit test files raced on shared process.env under CI: bun
runs all *.test.ts files concurrently in ONE process, and both files opted into
emit per-test by deleting FLIGHTDECK_EMIT_DISABLED / setting FLIGHTDECK_EVENTS_PATH,
restoring in afterEach. One file's afterEach restore clobbered the other mid-test
→ emit went no-op / wrote to the wrong buffer → the read threw ENOENT (25 CI-only
failures, deterministic under ubuntu-latest scheduling, invisible locally).

Merge both files into a single non-*.test.ts suite that default `bun test` does
NOT discover, and run it alone in its own bun process from validate.sh — no
concurrent FD file to race the shared env. All 48 original test cases preserved
(both files' full coverage, incl. the commutativity partial-probe regression).

Changes:
- tests/flightdeck-emit.suite.ts: merged flightdeck_emit.test.ts +
  handler_emit.test.ts; per-file hooks/state block-scoped in two describe blocks.
- tests/{flightdeck_emit,handler_emit}.test.ts: removed (moved into the suite).
- scripts/ci/validate.sh: after the main `bun test`, run the FD suite isolated
  via `bun test ./tests/flightdeck-emit.suite.ts` (leading ./ forces path, not
  name-filter); fails the script on failure.
- bunfig.toml: update the preload comment to reference the merged suite.

No product/emit logic changed.
